+++ dbt_dry_run/cli.py
@@ -3,11 +3,12 @@ from typing import Optional
 
 import typer
 from dbt.flags import DEFAULT_PROFILES_DIR
-from typer import Argument, Option
+from typer import Option
 
 from dbt_dry_run.adapter.service import DbtArgs, ProjectService
 from dbt_dry_run.exception import ManifestValidationError
 from dbt_dry_run.execution import dry_run_manifest
+from dbt_dry_run.flags import Flags, set_flags
 from dbt_dry_run.result_reporter import ResultReporter
 
 app = typer.Typer()
@@ -20,7 +21,9 @@ def dry_run(
     verbose: bool = False,
     report_path: Optional[str] = None,
     cli_vars: str = "{}",
+    skip_not_compiled: bool = False,
 ) -> int:
+    set_flags(Flags(skip_not_compiled=skip_not_compiled))
     args = DbtArgs(
         project_dir=project_dir,
         profiles_dir=os.path.abspath(profiles_dir),
@@ -47,13 +50,15 @@ def dry_run(
     return exit_code
 
 
+_SKIP_NOT_COMPILED_HELP = """
+    Whether or not the dry run should ignore models that are not compiled. This has several caveats that make this 
+    not a recommended option. The dbt manifest should generally be compiled with `--select *` to ensure good 
+    coverage
+"""
+
+
 @app.command()
 def run(
-    profile: Optional[str] = Argument(
-        None,
-        hidden=True,
-        help="Legacy parameter. You should not use this anymore see CHANGES.md in the github repo for how to migrate",
-    ),
     profiles_dir: str = Option(
         DEFAULT_PROFILES_DIR, help="[dbt] Where to search for `profiles.yml`"
     ),
@@ -64,13 +69,13 @@ def run(
     target: Optional[str] = Option(None, help="[dbt] Target profile"),
     verbose: bool = Option(False, help="Output verbose error messages"),
     report_path: Optional[str] = Option(None, help="Json path to dump report to"),
+    skip_not_compiled: bool = Option(
+        False, "--skip-not-compiled", help=_SKIP_NOT_COMPILED_HELP
+    ),
 ) -> None:
-    if profile is not None:
-        print(
-            "CLI format has changed see CHANGES.md v0.4.0 for instructions on how to migrate"
-        )
-        raise typer.Exit(1)
-    exit_code = dry_run(project_dir, profiles_dir, target, verbose, report_path, vars)
+    exit_code = dry_run(
+        project_dir, profiles_dir, target, verbose, report_path, vars, skip_not_compiled
+    )
     if exit_code > 0:
         raise typer.Exit(exit_code)