+++ dbt_dry_run/test/node_runner/test_model_runner.py
@@ -1,6 +1,12 @@
 from unittest.mock import MagicMock
 
-from dbt_dry_run.exception import SchemaChangeException, UpstreamFailedException
+from dbt_dry_run import flags
+from dbt_dry_run.exception import (
+    NotCompiledException,
+    SchemaChangeException,
+    UpstreamFailedException,
+)
+from dbt_dry_run.flags import Flags
 from dbt_dry_run.literals import enable_test_example_values
 from dbt_dry_run.models import BigQueryFieldType, Table, TableField
 from dbt_dry_run.models.manifest import NodeConfig, PartitionBy
@@ -460,3 +466,41 @@ def test_model_with_sql_header_executes_header_first() -> None:
     executed_sql = get_executed_sql(mock_sql_runner)
     assert executed_sql.startswith(pre_header_value)
     assert node.compiled_code in executed_sql
+
+
+def test_validate_node_fails_if_skip_not_compiled_is_false(
+    default_flags: Flags,
+) -> None:
+    flags.set_flags(flags.Flags(skip_not_compiled=False))
+    mock_sql_runner = MagicMock()
+    results = MagicMock()
+
+    node = SimpleNode(
+        unique_id="node1", depends_on=[], resource_type=ManifestScheduler.MODEL
+    ).to_node()
+    node.compiled = False
+
+    model_runner = ModelRunner(mock_sql_runner, results)
+
+    validation_result = model_runner.validate_node(node)
+    assert validation_result
+    assert validation_result.status == DryRunStatus.FAILURE
+    assert isinstance(validation_result.exception, NotCompiledException)
+
+
+def test_validate_node_skips_if_skip_not_compiled_is_true(default_flags: Flags) -> None:
+    flags.set_flags(flags.Flags(skip_not_compiled=True))
+    mock_sql_runner = MagicMock()
+    results = MagicMock()
+
+    node = SimpleNode(
+        unique_id="node1", depends_on=[], resource_type=ManifestScheduler.MODEL
+    ).to_node()
+    node.compiled = False
+
+    model_runner = ModelRunner(mock_sql_runner, results)
+
+    validation_result = model_runner.validate_node(node)
+    assert validation_result
+    assert validation_result.status == DryRunStatus.SKIPPED
+    assert validation_result.exception is None