+++ integration/projects/test_partially_compiled_project/models/run_mart_model.sql
@@ -1,0 +1,6 @@
+{{
+    config(alias="run_mart_model")
+}}
+
+SELECT *
+FROM {{ ref("run_staging_model") }}